Added the RadioHead library.
Optimized the module selection process in INO file.

Redefined the pinout in the code to match the planned rewiring of both
        the Simon and Alarm Module.

Created the RFHandler class to be a wrapper for the RadioHead library.

+++ LCD.h
@@ -1,13 +1,12 @@
 #ifndef __LCD_h_
 #define __LCD_h_
 
-#include "Arduino.h"
 #include "LiquidCrystal_I2C.h"
 #include <stdint.h>
 
 class LCD {
     private:
-        byte address;
+        int address;
         LiquidCrystal_I2C* lcd;
 
     public:
